Dissolved Base class for batch and bulk request options
