+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosBulkItemRequestOptions.java
@@ -4,7 +4,7 @@
 package com.azure.cosmos.models;
 
 import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.cosmos.CosmosItemOperation;
+import com.azure.cosmos.implementation.RequestOptions;
 import com.azure.cosmos.util.Beta;
 
 /**
@@ -12,8 +12,24 @@ import com.azure.cosmos.util.Beta;
  * creating bulk request using {@link CosmosBulkOperations}.
  */
 @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-public final class CosmosBulkItemRequestOptions
-        extends CosmosBulkItemRequestOptionsBase {
+public final class CosmosBulkItemRequestOptions {
+
+    private String ifMatchETag;
+    private String ifNoneMatchETag;
+    private Boolean contentResponseOnWriteEnabled;
+
+    public CosmosBulkItemRequestOptions() {
+    }
+
+    /**
+     * Gets the If-Match (ETag) associated with the operation in {@link CosmosItemOperation}.
+     *
+     * @return ifMatchETag the ifMatchETag associated with the request.
+     */
+    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public String getIfMatchETag() {
+        return this.ifMatchETag;
+    }
 
     /**
      * Sets the If-Match (ETag) associated with the operation in {@link CosmosItemOperation}.
@@ -23,8 +39,18 @@ public final class CosmosBulkItemRequestOptions
      */
     @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public CosmosBulkItemRequestOptions setIfMatchETag(final String ifMatchETag) {
-        super.setIfMatchETagCore(ifMatchETag);
+        this.ifMatchETag = ifMatchETag;
         return this;
+    }
+
+    /**
+     * Gets the If-None-Match (ETag) associated with the request in operation in {@link CosmosItemOperation}.
+     *
+     * @return the ifNoneMatchETag associated with the request.
+     */
+    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public String getIfNoneMatchETag() {
+        return this.ifNoneMatchETag;
     }
 
     /**
@@ -35,7 +61,7 @@ public final class CosmosBulkItemRequestOptions
      */
     @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public CosmosBulkItemRequestOptions setIfNoneMatchETag(final String ifNoneMatchEtag) {
-        super.setIfNoneMatchETagCore(ifNoneMatchEtag);
+        this.ifNoneMatchETag = ifNoneMatchEtag;
         return this;
     }
 
@@ -60,7 +86,33 @@ public final class CosmosBulkItemRequestOptions
      */
     @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
     public CosmosBulkItemRequestOptions setContentResponseOnWriteEnabled(Boolean contentResponseOnWriteEnabled) {
-        super.setContentResponseOnWriteEnabledCore(contentResponseOnWriteEnabled);
+        this.contentResponseOnWriteEnabled = contentResponseOnWriteEnabled;
         return this;
+    }
+
+    /**
+     * Gets the boolean to only return the headers and status code in Cosmos DB response
+     * in case of Create, Update and Delete operations in {@link CosmosItemOperation}.
+     *
+     * If set to false, service doesn't return payload in the response. It reduces networking
+     * and CPU load by not sending the payload back over the network and serializing it on the client.
+     *
+     * This feature does not impact RU usage for read or write operations.
+     *
+     * By-default, this is null.
+     *
+     * @return a boolean indicating whether payload will be included in the response or not for this operation.
+     */
+    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
+    public Boolean isContentResponseOnWriteEnabled() {
+        return this.contentResponseOnWriteEnabled;
+    }
+
+    RequestOptions toRequestOptions() {
+        final RequestOptions requestOptions = new RequestOptions();
+        requestOptions.setIfMatchETag(this.ifMatchETag);
+        requestOptions.setIfNoneMatchETag(this.ifNoneMatchETag);
+        requestOptions.setContentResponseOnWriteEnabled(this.contentResponseOnWriteEnabled);
+        return requestOptions;
     }
 }
