--- sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/CosmosBulkItemRequestOptionsBase.java
@@ -1,115 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.cosmos.models;
-
-import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.cosmos.CosmosItemOperation;
-import com.azure.cosmos.implementation.RequestOptions;
-import com.azure.cosmos.util.Beta;
-
-/**
- * Encapsulates options that can be specified for an operation used in Bulk execution. It can be passed while
- * creating bulk request using {@link CosmosBulkOperations}.
- */
-@Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-class CosmosBulkItemRequestOptionsBase {
-
-    private String ifMatchETag;
-    private String ifNoneMatchETag;
-    private Boolean contentResponseOnWriteEnabled;
-
-    protected CosmosBulkItemRequestOptionsBase(){
-    }
-
-    /**
-     * Gets the If-Match (ETag) associated with the operation in {@link CosmosItemOperation}.
-     *
-     * @return ifMatchETag the ifMatchETag associated with the request.
-     */
-    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public String getIfMatchETag() {
-        return this.ifMatchETag;
-    }
-
-    /**
-     * Sets the If-Match (ETag) associated with the operation in {@link CosmosItemOperation}.
-     *
-     * @param ifMatchETag the ifMatchETag associated with the request.
-     */
-    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public void setIfMatchETagCore(final String ifMatchETag){
-        this.ifMatchETag = ifMatchETag;
-        return;
-    }
-
-    /**
-     * Gets the If-None-Match (ETag) associated with the request in operation in {@link CosmosItemOperation}.
-     *
-     * @return the ifNoneMatchETag associated with the request.
-     */
-    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public String getIfNoneMatchETag() {
-        return this.ifNoneMatchETag;
-    }
-
-    /**
-     * Sets the If-None-Match (ETag) associated with the request in operation in {@link CosmosItemOperation}.
-     *
-     * @param ifNoneMatchEtag the ifNoneMatchETag associated with the request.
-     */
-    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public void setIfNoneMatchETagCore(final String ifNoneMatchEtag){
-        this.ifNoneMatchETag = ifNoneMatchEtag;
-        return;
-    }
-
-    /**
-     * Gets the boolean to only return the headers and status code in Cosmos DB response
-     * in case of Create, Update and Delete operations in {@link CosmosItemOperation}.
-     *
-     * If set to false, service doesn't return payload in the response. It reduces networking
-     * and CPU load by not sending the payload back over the network and serializing it on the client.
-     *
-     * This feature does not impact RU usage for read or write operations.
-     *
-     * By-default, this is null.
-     *
-     * @return a boolean indicating whether payload will be included in the response or not for this operation.
-     */
-    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public Boolean isContentResponseOnWriteEnabled() {
-        return contentResponseOnWriteEnabled;
-    }
-
-    /**
-     * Sets the boolean to only return the headers and status code in Cosmos DB response
-     * in case of Create, Update and Delete operations in {@link CosmosItemOperation}.
-     *
-     * If set to false, service doesn't return payload in the response. It reduces networking
-     * and CPU load by not sending the payload back over the network and serializing it on the client.
-     *
-     * This feature does not impact RU usage for read or write operations.
-     *
-     * By-default, this is null.
-     *
-     * NOTE: This flag is also present on {@link CosmosClientBuilder}, however if specified
-     * here, it will override the value specified in {@link CosmosClientBuilder} for this request.
-     *
-     * @param contentResponseOnWriteEnabled a boolean indicating whether payload will be included
-     * in the response or not for this operation.
-     */
-    @Beta(value = Beta.SinceVersion.V4_19_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public void setContentResponseOnWriteEnabledCore(Boolean contentResponseOnWriteEnabled){
-        this.contentResponseOnWriteEnabled = contentResponseOnWriteEnabled;
-        return;
-    }
-
-    RequestOptions toRequestOptions() {
-        final RequestOptions requestOptions = new RequestOptions();
-        requestOptions.setIfMatchETag(getIfMatchETag());
-        requestOptions.setIfNoneMatchETag(getIfNoneMatchETag());
-        requestOptions.setContentResponseOnWriteEnabled(isContentResponseOnWriteEnabled());
-        return requestOptions;
-    }
-}
